# api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/shows/cinemahallid/{id}/movieid/{movieId}":{"post":{"tags":["show-controller"],"operationId":"addShow","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"movieId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Show"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Show"}}}}}}},"/bookings":{"post":{"tags":["booking-controller"],"operationId":"makeBooking","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookingDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Booking"}}}}}}}},"/users":{"get":{"tags":["users-controller"],"operationId":"getUser","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Users"}}}}}}}},"/users/{id}":{"get":{"tags":["users-controller"],"operationId":"getUserById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Users"}}}}}}},"/users/{id}/bookings":{"get":{"tags":["booking-controller"],"operationId":"getBookings","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Booking"}}}}}}}},"/shows":{"get":{"tags":["show-controller"],"operationId":"getAllShow","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Show"}}}}}}}},"/movies/{id}":{"get":{"tags":["movie-controller"],"operationId":"getMovie","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Movie"}}}}}}},"/movies/search":{"get":{"tags":["movie-controller"],"operationId":"getMovieByFilter","parameters":[{"name":"title","in":"query","required":false,"schema":{"type":"string"}},{"name":"language","in":"query","required":false,"schema":{"type":"string"}},{"name":"releasedate","in":"query","required":false,"schema":{"type":"string","format":"date-time"}},{"name":"city","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Movie"}}}}}}}},"/cinemaseats/cinemahallid/{id}":{"get":{"tags":["cinema-seat-controller"],"operationId":"getCinemaSeats","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CinemaSeatReturn"}}}}}}},"/cinemas/{name}/cities":{"get":{"tags":["city-controller"],"operationId":"getCitiesByCinemaId","parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/City"}}}}}}}},"/bookings/{id}":{"get":{"tags":["booking-controller"],"operationId":"getBookingsById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Booking"}}}}}}}},"components":{"schemas":{"Show":{"type":"object","properties":{"showId":{"type":"integer","format":"int32"},"date":{"type":"string","format":"date-time"},"startTime":{"type":"string","format":"date-time"},"endTime":{"type":"string","format":"date-time"}}},"BookingDTO":{"type":"object","properties":{"cinema_seat_id":{"type":"array","items":{"type":"integer","format":"int32"}},"show_id":{"type":"integer","format":"int32"},"user_id":{"type":"integer","format":"int32"}}},"Booking":{"type":"object","properties":{"bookingId":{"type":"integer","format":"int32"},"time":{"type":"string","format":"date-time"},"status":{"type":"string"}}},"Users":{"required":["name","password","phone"],"type":"object","properties":{"userId":{"type":"integer","format":"int32"},"name":{"type":"string"},"password":{"type":"string"},"email":{"type":"string"},"phone":{"maximum":10,"minimum":10,"type":"integer","format":"int64"},"bookingList":{"type":"array","items":{"$ref":"#/components/schemas/Booking"}}}},"Movie":{"type":"object","properties":{"movieId":{"type":"integer","format":"int32"},"title":{"type":"string"},"duration":{"type":"string","format":"date-time"},"language":{"type":"string"},"releaseDate":{"type":"string","format":"date-time"},"country":{"type":"string"},"genre":{"type":"string"},"showList":{"type":"array","items":{"$ref":"#/components/schemas/Show"}},"description":{"type":"string"}}},"CinemaSeat":{"type":"object","properties":{"cinemaSeatId":{"type":"integer","format":"int32"},"seatNumber":{"type":"integer","format":"int32"},"type":{"type":"string"},"showSeatList":{"type":"array","items":{"$ref":"#/components/schemas/ShowSeat"}}}},"CinemaSeatReturn":{"type":"object","properties":{"bookedSeatList":{"type":"array","items":{"$ref":"#/components/schemas/CinemaSeat"}},"unBookedseatList":{"type":"array","items":{"$ref":"#/components/schemas/CinemaSeat"}}}},"ShowSeat":{"type":"object","properties":{"showSeatId":{"type":"integer","format":"int32"},"status":{"type":"string"},"price":{"type":"integer","format":"int32"}}},"City":{"type":"object","properties":{"cityId":{"type":"integer","format":"int32"},"name":{"type":"string"},"state":{"type":"string"}}}}}}